Completely reorganize source code
