+++ src/lib/MainWindow/WaitingSpinnerOverlay.py
